+++ packages/xgplayer-flv/src/plugin.js
@@ -1,6 +1,6 @@
 import { BasePlugin, Events, Errors } from 'xgplayer'
 import { EVENT } from 'xgplayer-streaming-shared'
-import { Flv } from './flv'
+import { Flv, logger } from './flv'
 import PluginExtension from './plugin-extension'
 
 export class FlvPlugin extends BasePlugin {
@@ -10,11 +10,14 @@ export class FlvPlugin extends BasePlugin {
     return 'flv'
   }
 
+  logger = logger
+
   /** @type {Flv} */
   flv = null;
 
   /** @type {PluginExtension} */
   pluginExtension = null
+
 
   /** @type {Flv} */
   get core () {
@@ -36,8 +39,13 @@ export class FlvPlugin extends BasePlugin {
     return this.flv?.loader
   }
 
+  get transferCost () {
+    return this.flv._transferCost.transferCost
+  }
+
   beforePlayerInit () {
     const config = this.player.config
+    const mediaElem = this.player.media || this.player.video
 
     if (!config.url) return
 
@@ -53,8 +61,8 @@ export class FlvPlugin extends BasePlugin {
     this.flv = new Flv({
       softDecode: this.softDecode,
       isLive: config.isLive,
-      media: this.player.video,
-      preProcessUrl: (url, ext) => this.player.preProcessUrl?.(url, ext) || {url, ext},
+      media: mediaElem,
+      preProcessUrl: (url, ext) => this.player?.preProcessUrl?.(url, ext) || {url, ext},
       ...flvOpts
     })
 
@@ -72,6 +80,7 @@ export class FlvPlugin extends BasePlugin {
     if (this.softDecode) {
       this.pluginExtension = new PluginExtension({
         media: this.player.video,
+        isLive: config.isLive,
         ...config.flv
       }, this)
       this.player.forceDegradeToVideo = (...args) => this.pluginExtension?.forceDegradeToVideo(...args)
@@ -92,6 +101,7 @@ export class FlvPlugin extends BasePlugin {
     this._transCoreEvent(EVENT.LOAD_RETRY)
     this._transCoreEvent(EVENT.SOURCEBUFFER_CREATED)
     this._transCoreEvent(EVENT.ANALYZE_DURATION_EXCEEDED)
+    this._transCoreEvent(EVENT.APPEND_BUFFER)
     this._transCoreEvent(EVENT.REMOVE_BUFFER)
     this._transCoreEvent(EVENT.BUFFEREOS)
     this._transCoreEvent(EVENT.KEYFRAME)