+++ packages/xgplayer-hls/src/hls/manifest-loader/parser/media.js
@@ -1,5 +1,5 @@
 import { MediaPlaylist, MediaSegment, MediaSegmentKey } from './model'
-import { getAbsoluteUrl, parseAttr, parseTag } from './utils'
+import { getAbsoluteUrl, parseAttr, parseTag, isValidDaterange } from './utils'
 
 export function parseMediaPlaylist (lines, parentUrl, useLowLatency) {
   const media = new MediaPlaylist()
@@ -63,7 +63,7 @@ export function parseMediaPlaylist (lines, parentUrl, useLowLatency) {
         media.canBlockReload = attr['CAN-BLOCK-RELOAD'] === 'YES'
         media.partHoldBack = parseFloat(attr['PART-HOLD-BACK'] || 0)
         media.canSkipUntil = parseFloat(attr['CAN-SKIP-UNTIL'] || 0)
-        media.canSkipDateRanges = attr['CAN-SKIP-DATERANGES'] === 'YES'
+        media.canSkipDateRanges = media.canSkipUntil > 0 && (attr['CAN-SKIP-DATERANGES'] === 'YES')
       }
         break
       case 'ENDLIST': {
@@ -102,7 +102,15 @@ export function parseMediaPlaylist (lines, parentUrl, useLowLatency) {
       }
 
         break
-      case 'PRELOAD-HINT':
+      case 'PRELOAD-HINT': {
+        const attr = parseAttr(data)
+        media.preloadHint = attr
+        if (attr['TYPE'] === 'PART' && attr['URI']) {
+          const tmp = attr['URI'].split('.ts')[0].split('-')
+          media.nextSN = tmp[3]
+          media.nextIndex = tmp[tmp.length - 1]
+        }
+      }
         break
       case 'PROGRAM-DATE-TIME':
         curSegment.dataTime = data
@@ -156,6 +164,28 @@ export function parseMediaPlaylist (lines, parentUrl, useLowLatency) {
         curSegment = new MediaSegment(parentUrl)
       }
         break
+      case 'SKIP': {
+        const attr = parseAttr(data)
+        const skippedSegments = parseInt(attr['SKIPPED-SEGMENTS'], 10)
+        if (skippedSegments <= Number.MAX_SAFE_INTEGER) {
+          media.skippedSegments += skippedSegments
+          curSN += skippedSegments
+        }
+      }
+        break
+      case 'DATERANGE': {
+        const attr = parseAttr(data)
+        const dateRangeWithSameId = media.dateRanges[attr.ID]
+        attr._startDate = dateRangeWithSameId ? dateRangeWithSameId._startDate : new Date(attr['START-DATE'])
+        const endDate = dateRangeWithSameId?._endDate || new Date(attr.END_DATE)
+        if (Number.isFinite(endDate)) {
+          attr._endDate = endDate
+        }
+        if (isValidDaterange(attr, dateRangeWithSameId) || media.skippedSegments) {
+          media.dateRanges[attr.ID] = attr
+        }
+      }
+        break
       default:
     }
   }
@@ -166,7 +196,6 @@ export function parseMediaPlaylist (lines, parentUrl, useLowLatency) {
   if (lastSegment) {
     if (endOfList) {
       lastSegment.isLast = true
-    
     }
     media.endSN = lastSegment.sn
     media.endPartIndex = lastSegment.partIndex