+++ packages/xgplayer-transmuxer/src/model/audio-track.js
@@ -12,6 +12,8 @@ export class AudioTrack {
 
   codec = ''
 
+  container = ''
+
   sequenceNumber = 0
 
   sampleDuration = 0
@@ -74,7 +76,15 @@ export class AudioTrack {
    * @returns {boolean}
    */
   exist () {
-    return !!(this.sampleRate && this.channelCount && this.codec && this.codecType === AudioCodecType.AAC)
+    return !!(
+      this.sampleRate &&
+      this.channelCount &&
+      (this.codec || this.container) &&
+      (this.codecType === AudioCodecType.AAC ||
+        this.codecType === AudioCodecType.G711PCMA ||
+        this.codecType === AudioCodecType.G711PCMU ||
+        this.codecType === AudioCodecType.OPUS || this.codecType === AudioCodecType.MP3)
+    )
   }
 
   /**