+++ packages/xgplayer-transmuxer/src/model/video-track.js
@@ -62,6 +62,8 @@ export class VideoTrack {
 
   isVideo = true
 
+  lastKeyFrameDts = 0
+
   kid = null
 
   pssh = null
@@ -111,6 +113,9 @@ export class VideoTrack {
    * @returns {boolean}
    */
   exist () {
+    if (/av01/.test(this.codec)) {
+      return true
+    }
     return !!(this.pps.length && this.sps.length && this.codec)
   }